backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.13.3
youtube-transcript-api==0.6.2
pytube==15.0.0
requests==2.31.0
python-multipart==0.0.9
cors==1.0.1
